Get Data From the ABARES Trade Dashboard

Fetches and imports ABARES trade data.

Usage

get_abares_trade(cache = TRUE)

Arguments

cache

Boolean Cache the ABARES trade data after download using tools::R_user_dir() to identify the proper directory for storing user data in a cache for this package. Defaults to TRUE, caching the files locally as a native R object. If FALSE, this function uses tempdir() and the files are deleted upon closing of the R session.

Value

A data.table object of the ABARES trade data

Note

Columns are renamed for consistency.

References

https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0

See also

Examples

trade <- get_abares_trade()

trade
#>           Fiscal_year Month Year_month Calendar_year Trade_code
#>                <char> <int>     <Date>         <int>      <i64>
#>        1:     1999-00    10 1999-01-01          1999   20089216
#>        2:     1999-00    12 1999-12-01          1999   20089216
#>        3:     1994-95     7 1994-07-01          1994   20085000
#>        4:     1994-95     1 1995-01-01          1995   20085000
#>        5:     1996-97    11 1996-11-01          1996   20085000
#>       ---                                                      
#> 10684802:     2022-23     3 2023-03-01          2023 2521000025
#> 10684803:     2022-23     5 2023-05-01          2023 2521000025
#> 10684804:     2023-24     9 2023-09-01          2023 2521000025
#> 10684805:     2023-24    12 2023-12-01          2023 2521000025
#> 10684806:     2023-24     3 2024-03-01          2024 2521000025
#>                              Overseas_location           State Australian_port
#>                                         <char>          <char>          <char>
#>        1:                               Canada South Australia        Adelaide
#>        2:                               Canada South Australia        Adelaide
#>        3: France (includes Andorra and Monaco) South Australia        Adelaide
#>        4: France (includes Andorra and Monaco) South Australia        Adelaide
#>        5: France (includes Andorra and Monaco) South Australia        Adelaide
#>       ---                                                                     
#> 10684802:                                Japan South Australia         Whyalla
#> 10684803:                                Japan South Australia         Whyalla
#> 10684804:                                Japan South Australia         Whyalla
#> 10684805:                                Japan South Australia         Whyalla
#> 10684806:                                Japan South Australia         Whyalla
#>               Unit Trade_flow Mode_of_transport    Value Quantity
#>             <char>     <char>            <char>    <num>    <num>
#>        1: B Carton    Exports               AIR   1177.0   414.00
#>        2: B Carton    Exports               AIR   1829.0   202.00
#>        3: B Carton    Exports               AIR   1020.0    25.00
#>        4: B Carton    Exports               AIR   1509.0    37.00
#>        5: B Carton    Exports               AIR   8883.0   150.00
#>       ---                                                        
#> 10684802:   Tonnes    Imports               SEA 842441.3 10467.71
#> 10684803:   Tonnes    Imports               SEA 756946.3 40000.00
#> 10684804:   Tonnes    Imports               SEA 872098.3 47300.00
#> 10684805:   Tonnes    Imports               SEA 833927.2 47100.00
#> 10684806:   Tonnes    Imports               SEA 710253.6  8825.22
#>           Confidentiality_flag
#>                          <int>
#>        1:                    0
#>        2:                    0
#>        3:                    0
#>        4:                    0
#>        5:                    0
#>       ---                     
#> 10684802:                    0
#> 10684803:                    0
#> 10684804:                    0
#> 10684805:                    0
#> 10684806:                    0