Fetch Australian Gridded Farm Data

Downloads The Australian Gridded Farm Data (AGFD) data and unzips the compressed files to NetCDF for importing.

Usage

get_agfd(fixed_prices = TRUE, cache = TRUE)

Arguments

fixed_prices

Boolean Download historical climate and prices or historical climate and fixed prices as described in (Hughes et al. 2022). Defaults to TRUE and downloads the data with historical climate and fixed prices “to isolate the effects of climate variability on financial incomes for broadacre farm businesses” (ABARES 2024). Using TRUE will download simulations where global output and input price indexes are fixed at values from the most recently completed financial year.

cache

Boolean Cache the Australian Gridded Farm Data files after download using tools::R_user_dir to identify the proper directory for storing user data in a cache for this package. Defaults to TRUE, caching the files locally. If FALSE, this function uses tempdir() and the files are deleted upon closing of the R session.

Value

A read.abares.agfd.nc.files object, a list of NetCDF files containing the Australian Gridded Farm Data

Details

From the ABARES website: “The Australian Gridded Farm Data (AGFD) are a set of national scale maps containing simulated data on historical broadacre farm business outcomes including farm profitability on an 0.05-degree (approximately 5 km) grid.
These data have been produced by ABARES as part of the ongoing Australian Agricultural Drought Indicator (AADI) project (previously known as the Drought Early Warning System Project) and were derived using ABARES farmpredict model, which in turn is based on ABARES Agricultural and Grazing Industries Survey (AAGIS) data.
Australian Agricultural Drought Indicator (AADI) project (previously known as the Drought Early Warning System Project) and were derived using ABARES farmpredict model, which in turn is based on ABARES Agricultural and Grazing Industries Survey (AAGIS) data.
These maps provide estimates of farm business profit, revenue, costs and production by location (grid cell) and year for the period 1990-91 to 2022-23. The data do not include actual observed outcomes but rather model predicted outcomes for representative or ‘typical’ broadacre farm businesses at each location considering likely farm characteristics and prevailing weather conditions and commodity prices.”
– ABARES, 2024-11-25

Both sets of data are large in file size, i.e., >1GB, and will require time to download.

References

Australian gridded farm data, Australian Bureau of Agricultural and Resource Economics and Sciences, Canberra, July 2024, DOI: 10.25814/7n6z-ev41. CC BY 4.0.

N. Hughes, W.Y. Soh, C. Boult, K. Lawson, Defining drought from the perspective of Australian farmers, Climate Risk Management, Volume 35, 2022, 100420, ISSN 2212-0963, DOI: 10.1016/j.crm.2022.100420.

See also

Examples

if (FALSE) { # interactive()
get_agfd()
}