List the File Path to Users' Cache Directory

Check what files exist in your read.abares file cache. This function will always return full file names, i.e., the directory path is prepended. See the help file for list.files() for more on the full.names argument. If you wish to strip the file path and only return the directory or file names, use basename(). See examples for more.

Usage

inspect_cache(recursive = FALSE)

Arguments

recursive

Boolean value indicating whether or not to list files in subdirectories of the cache directory. Defaults to FALSE returning only the top-level directories contained in the cache directory.

Value

A list of directories or files in the cache

See also

Examples

# not run because cached files shouldn't exist on CRAN or testing envs
if (FALSE) { # \dontrun{
# list directories in cache only
inspect_cache()

# list directory names, stripping the file path
basename(inspect_cache)

# list all files in subdirectories of the cache
inspect_cache(recursive = TRUE)

# list all files in subdirectories, stripping the file path
basename(inspect_cache(recursive_true))
} # }