Displays PDF Metadata for ABARES' "National Land Use" (NLUM) Raster Files in a Native Viewer

Each National Land Use (NLUM) raster file comes with a PDF of metadata. This function will open and display that file using the native PDF viewer for any system with a graphical user interface and PDF viewer configured. If the file does not exist locally, it will be fetched and displayed.

Usage

view_nlum_metadata_pdf()

Value

An invisible NULL. Called for its side-effects, opens the system's native PDF viewer to display the requested metadata PDF document.

See also

Examples

if (FALSE) { # interactive()
view_nlum_metadata_pdf()
}