Get Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1 for Local Use

Get Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1 for Local Use

Usage

get_soil_thickness(cache = TRUE)

Source

https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip

Arguments

cache

Boolean Cache the soil thickness data files after download using tools::R_user_dir() to identify the proper directory for storing user data in a cache for this package. Defaults to TRUE, caching the files locally. If FALSE, this function uses tempdir() and the files are deleted upon closing of the R session.

A custom print method is provided that will print the metadata associated with these data. Examples are provided for interacting with the metadata directly.

Value

A read.abares.soil.thickness object, which is a named list with the file path of the resulting ESRI Grid file and text file of metadata

References

https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6

See also

Examples

if (FALSE) { # interactive()
x <- get_soil_thickness()

# View the metadata with pretty printing
x

# Extract the metadata as an object in your R session and use it with
# {pander}, useful for Markdown files

library(pander)
y <- x$metadata
pander(y)
}