read.abares: Download and Import Agricultural Data from the Australian Bureau of Agricultural and Resource Economics and Sciences (ABARES) and Australian Bureau of Statistics (ABS)

Download and import agricultural data from the Australian Bureau of Agricultural and Resource Economics and Sciences (ABARES) https://www.agriculture.gov.au/abares and Australian Bureau of Statistics (ABS) https://www.abs.gov.au. Data types serviced include spreadsheets, CSV files, geospatial data including shape files and geotiffs covering topics including broadacre crops, livestock, soil data, commodities and more.

See also

Author

Maintainer: Adam H. Sparks adamhsparks@gmail.com (ORCID)

Other contributors:

Curtin University (ROR) (http://www.curtin.edu.au/) [funder, copyright holder]

Grains Research and Development Corporation (ROR) (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]

Jacob jacob@wujciak.de (ORCID) (Assisted with troubleshooting formatting in documentation to display '<' and '>' properly) [contributor]